% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sebastian.R
\name{sebastian}
\alias{sebastian}
\title{Title of the function, should be clear}
\usage{
sebastian(x, y, z)
}
\arguments{
\item{x}{An integer}

\item{y}{An integer}

\item{z}{An integer}
}
\value{
The sum of \code{x} and \code{y} to the power of \code{z}
}
\description{
What does the function do
}
\examples{
sebastian(1,2,3)
}
